Refactor backend into a modular structure
This commit refactors the monolithic `backend/main.py` file into a more organized and modular structure, following your specifications. The new structure separates concerns into `api`, `core`, `models`, and `utils` modules.

Key changes:
- API endpoints are now organized by domain in the `backend/api/` directory.
- Core components like database, Redis, security, and middleware are in `backend/core/`.
- Pydantic models are defined in `backend/models/schemas.py`.
- Utility functions are located in `backend/utils/`.
- The main `backend/main.py` file is now a lightweight application assembler.
- Added comments throughout the new modules to improve code clarity.

Note: The existing tests in `backend/tests/` are broken by this refactoring and will need to be updated to reflect the new application structure.
